refactor: reorganize annotation-related types and update imports
